Fixes to the generics handling in the Fody weaver

+++ NetFabric.Hyperlinq.TestCollections/TestEnumerable.cs
@@ -6,40 +6,40 @@ namespace NetFabric.Hyperlinq
 {
     public static class TestEnumerable
     {
-        public static IEnumerable<long> ReferenceType(long count)
+        public static IEnumerable<int> ReferenceType(int count)
         {
             for (var value = 0; value < count; value++)
                 yield return value;            
         }
 
-        public static Enumerable ValueType(long count) 
+        public static Enumerable ValueType(int count) 
             => new Enumerable(count);
 
-        public readonly struct Enumerable : IEnumerable<long>
+        public readonly struct Enumerable : IEnumerable<int>
         {
-            readonly long count;
+            readonly int count;
 
-            public Enumerable(long count)
+            public Enumerable(int count)
             {
                 this.count = count;
             }
 
             public Enumerator GetEnumerator() => new Enumerator(count);
-            IEnumerator<long> IEnumerable<long>.GetEnumerator() => new Enumerator(count);
+            IEnumerator<int> IEnumerable<int>.GetEnumerator() => new Enumerator(count);
             IEnumerator IEnumerable.GetEnumerator() => new Enumerator(count);
 
-            public struct Enumerator : IEnumerator<long>
+            public struct Enumerator : IEnumerator<int>
             {
-                readonly long count;
-                long current;
+                readonly int count;
+                int current;
 
-                public Enumerator(long count)
+                public Enumerator(int count)
                 {
                     this.count = count;
                     current = -1;
                 }
 
-                public long Current => current;
+                public int Current => current;
                 object IEnumerator.Current => current;
 
                 public bool MoveNext() => ++current < count;
